+++ src/lib/get-auth.ts
@@ -44,7 +44,7 @@ function getKey(user: RMOAS.User, scheme: RMOAS.KeyedSecuritySchemeObject): auth
  */
 export function getByScheme(
   user: RMOAS.User,
-  scheme = <RMOAS.KeyedSecuritySchemeObject>{},
+  scheme = <RMOAS.KeyedSecuritySchemeObject>{}, // eslint-disable-line default-param-last
   selectedApp?: string | number
 ): authKey {
   if (user?.keys) {
@@ -74,7 +74,7 @@ export default function getAuth(
   user: RMOAS.User,
   selectedApp?: string | number
 ) {
-  return Object.keys(api.components.securitySchemes)
+  return Object.keys(api?.components?.securitySchemes || {})
     .map(scheme => {
       return {
         [scheme]: getByScheme(